reorganize small data classes to be in one file
