+++ simulariumio/data_objects/frame_data.py
@@ -1,5 +1,47 @@
-from typing import Dict, Union
-import numpy as np
+from typing import Dict, Union, Tuple
+
+
+class DataIndices:
+    offset: int
+    length: int
+
+    def __init__(self, offset: int, length: int):
+        """
+        This object holds offset and length data for a binary
+        simularium data block
+        """
+        self.offset = offset
+        self.length = length
+
+    def get_start_end_indices(self) -> Tuple[int, int]:
+        """
+        Return the start and end indicies for the data block
+        """
+        end = self.offset + self.length - 1
+        return self.offset, end
+
+
+class FrameMetadata:
+    offset: int
+    length: int
+    frame_number: int
+    time: float
+
+    def __init__(self, offset: int, length: int, frame_number: int, time: float):
+        """
+        This object holds metadata for a single frame of simularium data
+        """
+        self.offset = offset
+        self.length = length
+        self.frame_number = frame_number
+        self.time = time
+
+    def get_start_end_indices(self) -> Tuple[int, int]:
+        """
+        Return the start and end indicies for the frame
+        """
+        end = self.offset + self.length - 1
+        return self.offset, end
 
 
 class FrameData:
